requirements.txt
aiofiles==24.1.0
aiokafka==0.12.0
aiologger==0.7.0
anyio==4.9.0
arrow==1.3.0
async-timeout==5.0.1
asyncpg==0.30.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
discord-webhook==1.4.1
distro==1.9.0
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
motor==3.7.0
numpy==2.3.1
orjson==3.10.16
packaging==25.0
pandas==2.3.1
pyarrow==21.0.0
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
RapidFuzz==3.13.0
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
